refactor(sdk): rename queryweaver_sdk → queryweaver
Match the import path to the package name. After this, ``pip install queryweaver`` is followed by ``from queryweaver import QueryWeaver`` (the standard Python convention) instead of the awkward ``queryweaver_sdk`` mismatch.

git mv preserves history on the four module files. All importers updated by sed:
- api/core/text2sql.py, result_models.py, request_models.py
- examples/ecommerce_example.py + examples/README.md
- tests/test_sdk/conftest.py, test_queryweaver.py
- pyproject.toml ([tool.hatch.build.targets.wheel].packages and the sdist include list)
- README.md SDK section

Verified: pylint 10/10, 14/14 SDK tests pass, fresh-venv install + ``from queryweaver import QueryWeaver`` works, end-to-end example against demo postgres unchanged.
