+++ queryweaver/models.py
@@ -2,7 +2,7 @@
 
 Thin re-export shim. The dataclasses live in ``api.core.result_models`` /
 ``api.core.request_models`` so the server code and the SDK share a single
-source of truth. Importing from ``queryweaver_sdk.models`` still works for
+source of truth. Importing from ``queryweaver.models`` still works for
 external consumers.
 """
